Extracted the logic for parsing Maven lock files of rules_jvm_external into separate classes.
