+++ metals/src/main/scala/scala/meta/internal/metals/mbt/importer/BazelMavenJsonImporter.scala
@@ -1,6 +1,5 @@
 package scala.meta.internal.metals.mbt.importer
 
-import java.nio.file.Files
 import java.nio.file.Path
 
 import scala.collection.concurrent.TrieMap
@@ -8,12 +7,12 @@ import scala.util.Try
 
 import scala.meta.internal.metals.MetalsEnrichments._
 import scala.meta.internal.metals.mbt.MbtDependencyModule
+import scala.meta.internal.metals.mbt.importer.bazel.MavenLockFileParser
 import scala.meta.internal.semver.SemVer
 import scala.meta.io.AbsolutePath
 
 import com.google.gson.Gson
 import com.google.gson.JsonArray
-import com.google.gson.JsonElement
 import com.google.gson.JsonObject
 
 /**
@@ -28,7 +27,7 @@ object BazelMavenJsonImporter {
   private val gson = new Gson()
 
   /** Pre-scanned snapshot of an external directory listing. */
-  private case class ScannedExtDir(
+  private[importer] case class ScannedExtDir(
       dir: AbsolutePath,
       entries: List[AbsolutePath],
   )
@@ -521,27 +520,6 @@ object BazelMavenJsonImporter {
 
   /**
    * Extract artifacts from maven_install.json.
-   *
-   * Supports two formats:
-   *
-   * New format (rules_jvm_external v5+):
-   * {
-   *   "artifacts": {
-   *     "com.google.guava:guava": {
-   *       "version": "31.1-jre",
-   *       "shasums": { "jar": "...", "sources": "..." }
-   *     }
-   *   }
-   * }
-   *
-   * Legacy format (rules_jvm_external v4 and earlier):
-   * {
-   *   "dependency_tree": {
-   *     "dependencies": [
-   *       { "coord": "com.google.guava:guava:31.1-jre", "file": "v1/https/..." }
-   *     ]
-   *   }
-   * }
    */
   private def extractArtifacts(
       json: JsonObject,
@@ -557,13 +535,9 @@ object BazelMavenJsonImporter {
       externalDirs(projectDir, outputBase)
         .flatMap(d => Try(ScannedExtDir(d, d.list.toList)).toOption)
 
-    MavenLockFormat.of(json) match {
-      case MavenLockFormat.NewFormat(artifacts) =>
-        extractFromNewFormat(artifacts, repositoryNames, extDirs)
-      case MavenLockFormat.LegacyFormat(deps) =>
-        scribe.debug("Using legacy dependency_tree format")
-        extractFromLegacyFormat(deps, repositoryNames, extDirs)
-      case MavenLockFormat.Unknown =>
+    MavenLockFileParser.create(json) match {
+      case Some(parser) => parser.parse(repositoryNames, extDirs)
+      case None =>
         scribe.debug(
           s"No 'artifacts' or 'dependency_tree.dependencies' found in $lockFile"
         )
@@ -571,548 +545,4 @@ object BazelMavenJsonImporter {
     }
   }
 
-  /**
-   * Extract from new format: { "artifacts": { "group:artifact": { "version": "..." } } }
-   */
-  private def extractFromNewFormat(
-      artifacts: JsonObject,
-      repositoryNames: Seq[String],
-      extDirs: Seq[ScannedExtDir],
-  ): Seq[MbtDependencyModule] = {
-    artifacts.entrySet().asScala.toSeq.flatMap { entry =>
-      val coordKey = entry.getKey // e.g., "com.google.guava:guava"
-      val artifactInfo = entry.getValue
-      if (!artifactInfo.isJsonObject) None
-      else
-        parseArtifact(
-          coordKey,
-          artifactInfo,
-          repositoryNames,
-          extDirs,
-        )
-    }
-  }
-
-  /**
-   * Extract from legacy format: { "dependency_tree": { "dependencies": [...] } }
-   */
-  private def extractFromLegacyFormat(
-      dependencies: JsonArray,
-      repositoryNames: Seq[String],
-      extDirs: Seq[ScannedExtDir],
-  ): Seq[MbtDependencyModule] = {
-    val deps = dependencies.iterator().asScala.toSeq
-
-    // Group by base coordinate to pair JARs with sources
-    val jarDeps = deps.filter { elem =>
-      if (!elem.isJsonObject) false
-      else {
-        val coord = Option(elem.getAsJsonObject.get("coord"))
-          .filter(_.isJsonPrimitive)
-          .map(_.getAsString)
-          .getOrElse("")
-        !coord.contains(":jar:sources:") && !coord.contains("-sources.")
-      }
-    }
-
-    jarDeps.flatMap { elem =>
-      parseLegacyDependency(elem, deps, repositoryNames, extDirs)
-    }
-  }
-
-  private def parseLegacyDependency(
-      elem: JsonElement,
-      allDeps: Seq[JsonElement],
-      repositoryNames: Seq[String],
-      extDirs: Seq[ScannedExtDir],
-  ): Option[MbtDependencyModule] = {
-    Try {
-      val obj = elem.getAsJsonObject
-      val coord =
-        obj.get("coord").getAsString // e.g., "com.google.guava:guava:31.1-jre"
-      val file = Option(obj.get("file")).map(_.getAsString)
-      val url = Option(obj.get("url")).map(_.getAsString)
-
-      // Parse coordinate: group:artifact[:packaging[:classifier]]:version
-      val parts = coord.split(":")
-      if (parts.length < 3) None
-      else {
-        val (groupId, artifactId, version) = (parts(0), parts(1), parts.last)
-
-        if (version.contains("SNAPSHOT")) {
-          scribe.debug(s"Skipping SNAPSHOT: $coord")
-          None
-        } else {
-          val id = s"$groupId:$artifactId:$version"
-          val groupPath = groupId.replace('.', '/')
-          val jarName = s"$artifactId-$version.jar"
-
-          // Find JAR path - try local paths first
-          val jarPath = file
-            .flatMap { f =>
-              findJarFromLegacyPath(f, repositoryNames, extDirs)
-            }
-            .orElse {
-              findJarPath(
-                groupId,
-                artifactId,
-                version,
-                repositoryNames,
-                extDirs,
-              )
-            }
-            .orElse {
-              scribe.warn(s"Downloading from URL, could not find locally: $url")
-              url.flatMap { u =>
-                downloadToM2Cache(u, groupPath, artifactId, version, jarName)
-              }
-            }
-
-          jarPath.map { jar =>
-            // Look for corresponding sources in the dependencies array
-            val sourcesCoord = s"$groupId:$artifactId:jar:sources:$version"
-            val sourcesEntry = allDeps.find { e =>
-              e.isJsonObject &&
-              Option(e.getAsJsonObject.get("coord"))
-                .map(_.getAsString)
-                .contains(sourcesCoord)
-            }
-
-            val sourcesFile = sourcesEntry.flatMap { e =>
-              Option(e.getAsJsonObject.get("file")).map(_.getAsString)
-            }
-            val sourcesUrl = sourcesEntry.flatMap { e =>
-              Option(e.getAsJsonObject.get("url")).map(_.getAsString)
-            }
-
-            val sourcesJarName = s"$artifactId-$version-sources.jar"
-            val sourcesPath = sourcesFile
-              .flatMap { f =>
-                findJarFromLegacyPath(f, repositoryNames, extDirs)
-              }
-              .orElse {
-                findSourcesPath(
-                  groupId,
-                  artifactId,
-                  version,
-                  repositoryNames,
-                  extDirs,
-                )
-              }
-              .orElse {
-                // Download sources from URL if available
-                sourcesUrl.flatMap { u =>
-                  downloadToM2Cache(
-                    u,
-                    groupPath,
-                    artifactId,
-                    version,
-                    sourcesJarName,
-                  )
-                }
-              }
-
-            MbtDependencyModule(
-              id = id,
-              jar = jar,
-              sources = sourcesPath.orNull,
-            )
-          }
-        }
-      }
-    }.toOption.flatten
-  }
-
-  /**
-   * Download a JAR from URL to the local Maven cache.
-   */
-  private def downloadToM2Cache(
-      url: String,
-      groupPath: String,
-      artifactId: String,
-      version: String,
-      jarName: String,
-  ): Option[String] = {
-    Try {
-      val m2Dir = Path.of(
-        System.getProperty("user.home"),
-        ".m2",
-        "repository",
-        groupPath,
-        artifactId,
-        version,
-      )
-      val targetPath = m2Dir.resolve(jarName)
-
-      // Check if already exists
-      if (Files.exists(targetPath)) {
-        Some(targetPath.toUri.toString)
-      } else {
-        scribe.debug(s"Downloading: $url")
-
-        val urlLower = url.toLowerCase
-        if (
-          !urlLower.startsWith("http://") && !urlLower.startsWith("https://")
-        ) {
-          scribe.debug(
-            s"Skipping download: URL must use http or https, got: $url"
-          )
-          None
-        } else {
-
-          // Create directory
-          Files.createDirectories(m2Dir)
-
-          // Download file
-          val connection = new java.net.URL(url).openConnection()
-          connection.setConnectTimeout(10000)
-          connection.setReadTimeout(30000)
-
-          val in = connection.getInputStream
-          try {
-            Files.copy(in, targetPath)
-            scribe.debug(s"Downloaded to: $targetPath")
-            Some(targetPath.toUri.toString)
-          } finally {
-            in.close()
-          }
-        }
-      }
-    }.toOption.flatten
-  }
-
-  /**
-   * Find JAR from legacy format's "file" path.
-   * The file path is like: v1/https/repo1.maven.org/maven2/com/google/guava/guava/31.1-jre/guava-31.1-jre.jar
-   */
-  private def findJarFromLegacyPath(
-      filePath: String,
-      repositoryNames: Seq[String],
-      extDirs: Seq[ScannedExtDir],
-  ): Option[String] = {
-    val patterns =
-      (repositoryNames.map(repo => s"$repo/$filePath") ++
-        Seq(s"maven/$filePath", filePath)).distinct
-
-    extDirs.view
-      .flatMap(findInExternalRepositories(_, patterns, repositoryNames))
-      .headOption
-      .map(_.toURI.toString)
-  }
-
-  private def parseArtifact(
-      coordKey: String,
-      artifactInfo: JsonElement,
-      repositoryNames: Seq[String],
-      extDirs: Seq[ScannedExtDir],
-  ): Option[MbtDependencyModule] = {
-    val info = artifactInfo.getAsJsonObject
-
-    // Get version
-    val version = Option(info.get("version"))
-      .filter(_.isJsonPrimitive)
-      .map(_.getAsString)
-
-    version.flatMap { v =>
-      if (v.contains("SNAPSHOT")) {
-        scribe.debug(s"Skipping SNAPSHOT: $coordKey:$v")
-        None
-      } else {
-        val parts = coordKey.split(":")
-        if (parts.length != 2) None
-        else {
-          val (groupId, artifactId) = (parts(0), parts(1))
-          val id = s"$groupId:$artifactId:$v"
-
-          // Find JAR file path
-          val jarPath =
-            findJarPath(groupId, artifactId, v, repositoryNames, extDirs)
-
-          jarPath.map { jar =>
-            val sourcesPath =
-              findSourcesPath(groupId, artifactId, v, repositoryNames, extDirs)
-
-            MbtDependencyModule(
-              id = id,
-              jar = jar,
-              sources = sourcesPath.orNull,
-            )
-          }
-        }
-      }
-    }
-  }
-
-  private def workspaceArtifactPatterns(
-      repositoryNames: Seq[String],
-      groupPath: String,
-      artifactId: String,
-      version: String,
-      fileName: String,
-  ): Seq[String] =
-    (repositoryNames.map(repo =>
-      s"$repo/v1/https/repo1.maven.org/maven2/$groupPath/$artifactId/$version/$fileName"
-    ) ++ Seq(
-      s"maven/v1/https/repo1.maven.org/maven2/$groupPath/$artifactId/$version/$fileName",
-      s"maven/$groupPath/$artifactId/$version/$fileName",
-    )).distinct
-
-  /**
-   * Find the JAR file in the Bazel external directory.
-   *
-   * Bazel stores external dependencies in different patterns:
-   *
-   * Bzlmod mode (Bazel 6+):
-   *   {output_base}/external/rules_jvm_external~{version}~maven~{artifact_dir}_{version}/file/v1/{group_path}/{artifact}/{version}/{artifact}-{version}.jar
-   *   Example: rules_jvm_external~6.2~maven~com_google_guava_guava_32_1_3_android/file/v1/com/google/guava/guava/32.1.3-android/guava-32.1.3-android.jar
-   *
-   * WORKSPACE mode:
-   *   {output_base}/external/maven/v1/https/repo1.maven.org/maven2/{group_path}/{artifact}/{version}/{artifact}-{version}.jar
-   */
-  private def findJarPath(
-      groupId: String,
-      artifactId: String,
-      version: String,
-      repositoryNames: Seq[String],
-      extDirs: Seq[ScannedExtDir],
-  ): Option[String] = {
-    val groupPath = groupId.replace('.', '/')
-    val jarName = s"$artifactId-$version.jar"
-
-    val bzlmodArtifactDir = toBzlmodArtifactDir(groupId, artifactId, version)
-
-    val workspacePatterns =
-      workspaceArtifactPatterns(
-        repositoryNames,
-        groupPath,
-        artifactId,
-        version,
-        jarName,
-      )
-
-    extDirs.view
-      .flatMap { scanned =>
-        findBzlmodJar(
-          scanned,
-          bzlmodArtifactDir,
-          groupPath,
-          artifactId,
-          version,
-          jarName,
-        )
-          .orElse(
-            findInExternalRepositories(
-              scanned,
-              workspacePatterns,
-              repositoryNames,
-            )
-          )
-      }
-      .headOption
-      .orElse {
-        val m2Path = Path.of(
-          System.getProperty("user.home"),
-          ".m2",
-          "repository",
-          groupPath,
-          artifactId,
-          version,
-          jarName,
-        )
-        if (Files.exists(m2Path)) Some(AbsolutePath(m2Path)) else None
-      }
-      .map(_.toURI.toString)
-  }
-
-  /**
-   * Convert artifact coordinates to Bzlmod directory name format.
-   * Example: com.google.guava:guava:32.1.3-android -> com_google_guava_guava_32_1_3_android
-   */
-  private def toBzlmodArtifactDir(
-      groupId: String,
-      artifactId: String,
-      version: String,
-  ): String = {
-    val sanitize = (s: String) =>
-      s.replace('.', '_').replace('-', '_').replace(':', '_')
-    s"${sanitize(groupId)}_${sanitize(artifactId)}_${sanitize(version)}"
-  }
-
-  /**
-   * Find JAR in Bzlmod external directory by scanning for rules_jvm_external prefix.
-   *
-   * Supports multiple Bazel versions:
-   * - Bazel 7.x: rules_jvm_external~{version}~maven~{artifactDir}
-   * - Bazel 8.x: rules_jvm_external++maven+{artifactDir}
-   */
-  private def findBzlmodJar(
-      scanned: ScannedExtDir,
-      artifactDir: String,
-      groupPath: String,
-      artifactId: String,
-      version: String,
-      jarName: String,
-  ): Option[AbsolutePath] = {
-    // Look for directory matching Bazel 7.x pattern: rules_jvm_external~{version}~maven~{artifactDir}
-    val bazel7Dir = scanned.entries.find { entry =>
-      val name = entry.filename.toString
-      name.startsWith("rules_jvm_external~") &&
-      name.contains("~maven~") &&
-      name.endsWith(s"~$artifactDir")
-    }
-
-    // Look for directory matching Bazel 8.x pattern: rules_jvm_external++maven+{artifactDir}
-    val bazel8Dir = scanned.entries.find { entry =>
-      val name = entry.filename.toString
-      name.startsWith("rules_jvm_external++maven+") &&
-      name.endsWith(s"+$artifactDir")
-    }
-
-    val matchingDir = bazel7Dir.orElse(bazel8Dir)
-
-    matchingDir.flatMap { dir =>
-      val jarPath =
-        dir.resolve(s"file/v1/$groupPath/$artifactId/$version/$jarName")
-      if (jarPath.exists) {
-        scribe.debug(s"Found Bzlmod JAR: $jarPath")
-        Some(jarPath)
-      } else {
-        // Also try without the nested path (some artifacts are directly in the directory)
-        val altJarPath = dir.resolve(s"file/$jarName")
-        if (altJarPath.exists) {
-          scribe.debug(s"Found Bzlmod JAR (alt): $altJarPath")
-          Some(altJarPath)
-        } else {
-          None
-        }
-      }
-    }
-  }
-
-  /**
-   * Find JARs inside Bzlmod maven repository roots such as:
-   * rules_jvm_external++maven+unpinned_maven/v1/https/...
-   */
-  private def findInExternalRepositories(
-      scanned: ScannedExtDir,
-      relativePaths: Seq[String],
-      repositoryNames: Seq[String],
-  ): Option[AbsolutePath] = {
-    val directMatch =
-      relativePaths.map(path => scanned.dir.resolve(path)).find(_.exists)
-
-    directMatch.orElse {
-      val bzlmodRepositoryNames =
-        repositoryNames.flatMap(repo => Seq(repo, s"unpinned_$repo")).distinct
-
-      val candidateRepositories = scanned.entries.filter { entry =>
-        val name = entry.filename.toString
-        entry.isDirectory &&
-        name.startsWith("rules_jvm_external") &&
-        bzlmodRepositoryNames.exists { repo =>
-          name.endsWith(s"+$repo") || name.endsWith(s"~$repo")
-        }
-      }
-
-      val repositoryRelativePaths = relativePaths.flatMap { path =>
-        path +: bzlmodRepositoryNames.map(repo => path.stripPrefix(s"$repo/"))
-      }.distinct
-
-      val found = candidateRepositories.flatMap { repo =>
-        repositoryRelativePaths
-          .map(path => repo.resolve(path))
-          .find(_.exists)
-      }.headOption
-
-      if (candidateRepositories.nonEmpty && found.isEmpty) {
-        scribe.warn(
-          s"JAR not found in central maven repos " +
-            s"${candidateRepositories.map(_.filename).mkString(", ")}; " +
-            s"tried: [${repositoryRelativePaths.mkString(", ")}]"
-        )
-      }
-
-      found
-    }
-  }
-
-  /**
-   * Find the sources JAR file.
-   */
-  private def findSourcesPath(
-      groupId: String,
-      artifactId: String,
-      version: String,
-      repositoryNames: Seq[String],
-      extDirs: Seq[ScannedExtDir],
-  ): Option[String] = {
-    val groupPath = groupId.replace('.', '/')
-    val sourcesJarName = s"$artifactId-$version-sources.jar"
-
-    val bzlmodSourcesDirs = toBzlmodSourcesDirs(groupId, artifactId, version)
-
-    val workspacePatterns =
-      workspaceArtifactPatterns(
-        repositoryNames,
-        groupPath,
-        artifactId,
-        version,
-        sourcesJarName,
-      )
-
-    extDirs.view
-      .flatMap { scanned =>
-        bzlmodSourcesDirs.view
-          .flatMap(
-            findBzlmodJar(
-              scanned,
-              _,
-              groupPath,
-              artifactId,
-              version,
-              sourcesJarName,
-            )
-          )
-          .headOption
-          .orElse(
-            findInExternalRepositories(
-              scanned,
-              workspacePatterns,
-              repositoryNames,
-            )
-          )
-      }
-      .headOption
-      .orElse {
-        // Fall back to local Maven cache
-        val m2Path = Path.of(
-          System.getProperty("user.home"),
-          ".m2",
-          "repository",
-          groupPath,
-          artifactId,
-          version,
-          sourcesJarName,
-        )
-        if (Files.exists(m2Path)) Some(AbsolutePath(m2Path)) else None
-      }
-      .map(_.toURI.toString)
-  }
-
-  /**
-   * Convert artifact coordinates to Bzlmod sources directory name formats.
-   * Example: com.google.guava:guava:32.1.3-android -> com_google_guava_guava_sources_32_1_3_android
-   */
-  private def toBzlmodSourcesDirs(
-      groupId: String,
-      artifactId: String,
-      version: String,
-  ): Seq[String] = {
-    val sanitize = (s: String) =>
-      s.replace('.', '_').replace('-', '_').replace(':', '_')
-    Seq(
-      s"${sanitize(groupId)}_${sanitize(artifactId)}_sources_${sanitize(version)}",
-      s"${sanitize(groupId)}_${sanitize(artifactId)}_jar_sources_${sanitize(version)}",
-    )
-  }
 }
